Moved bonus parsing to a new file
