+++ lib/json/JsonRandom.cpp
@@ -13,7 +13,7 @@
 
 #include <vstd/StringUtils.h>
 
-#include "JsonUtils.h"
+#include "JsonBonus.h"
 
 #include "../CRandomGenerator.h"
 #include "../constants/StringConstants.h"
