+++ lib/serializer/JsonUpdater.cpp
@@ -12,7 +12,7 @@
 
 #include "../bonuses/CBonusSystemNode.h"
 #include "../bonuses/Bonus.h"
-#include "../json/JsonUtils.h"
+#include "../json/JsonBonus.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
